build: stop using meson wrap
The mix of submodules and meson wrap-promoted sub-subprojects don't play
well: when we update a submodule we want to get the latest sub-submodule
version, not the one we had promoted. Also when fixing something in our
copy of the sub-submodule we want it to reflect on the submodule status,
marking it as dirty.

Simply add a symlink to where we were getting the sub-subproject from.