Do not grab device by default
We currently don't have a way to trigger actions on the RC itself,
so we may resort to evtest for mundane tasks like switching it on/off,
changing the RC stack in use, etc.

For that to work we can't take the device exclusively, otherwise evtest
won't get any event.