bundler: add options for qemu and output
In some environments we may not have access to binfmt being registered
with the kernel. This happens on gitlab-ci while trying to execute the
ARM's ld. Just allow to receive the qemu binary as an optional argument.

Also, since we are adding options, add a -o option so user can tell
where to save the tarball. Differently from previous behavior, it will
append to an existing tarball rather than re-creating it.