distro: sk2: move files to create the overlay only
Organize files in a "overlay"-like manner rather than the absolute
directory. The place where binary, libraries and configuration files are
kept will be up to a future install script.

This makes it easier to reuse the same strategy for a different board.
E.g. in a normal Linux distro we would just "copy" the contents of
distro/<name>/ to /.